abstracted the calls to a hero service
we moved everything about the store to a store folder

created model folder

created a reducers file

removed code we were not using (extra selectors in the reducers file)
