+++ src/client/app/app.module.ts
@@ -2,16 +2,13 @@ import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { StoreModule, Store } from '@ngrx/store';
+import { EffectsModule } from '@ngrx/effects';
 
 import { AppComponent } from './app.component';
-import { HeroDataService } from './hero-data.service';
+import { HeroService, HeroDataService, HeroEffects, reducers } from './store';
 import { HeroListComponent } from './hero-list.component';
 import { HeroDetailComponent } from './hero-detail.component';
-import { StoreModule } from '@ngrx/store';
-import { EffectsModule } from '@ngrx/effects';
-import { HeroEffects } from './hero.effects';
-import { reducers } from './hero.reducer';
-import { HeroService } from './hero.service';
 
 @NgModule({
   declarations: [AppComponent, HeroListComponent, HeroDetailComponent],
@@ -20,10 +17,12 @@ import { HeroService } from './hero.service';
     FormsModule,
     HttpClientModule,
     StoreModule.forRoot(reducers),
-    // StoreModule.forRoot({ reducers }), // TODO: this is not working
     EffectsModule.forRoot([HeroEffects])
   ],
-  providers: [HeroDataService, HeroService],
+  providers: [
+    HeroDataService,
+    HeroService
+  ],
   bootstrap: [AppComponent]
 })
 export class AppModule {}
