+++ src/client/app/model/hero.ts
