+++ src/client/app/store/index.ts
@@ -1,0 +1,6 @@
+export * from './hero-data.service';
+export * from './hero.action';
+export * from './hero.effects';
+export * from './hero.reducer';
+export * from './hero.service';
+export * from './reducers';
